ml-service/requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
torch==2.2.1
torchvision==0.17.1
Pillow==10.2.0
requests==2.31.0
numpy==1.26.4
python-multipart==0.0.9
pydantic==2.6.3
python-dotenv==1.0.0
